Refactor: Extract helpers into separate modules
Move utility functions into dedicated modules for better organization and reusability.
