+++ ai_proxy/logdb/ingest.py
@@ -12,6 +12,27 @@ from typing import Dict, Iterator, List, Optional, Tuple
 from .partitioning import ensure_partition_database, ensure_control_database
 from .schema import open_connection_with_pragmas
 
+# Extracted helpers
+from .parsers.log_parser import (
+    _safe_iso_to_datetime,
+    _iter_json_blocks,
+    _parse_log_entry,
+    _normalize_entry,
+    _compute_request_id,
+)
+from .utils.file_utils import (
+    _derive_server_id,
+    _file_sha256,
+    _file_prefix_sha256,
+    _env_int,
+)
+from .utils.checkpoint import (
+    _ensure_servers_row,
+    _upsert_ingest_checkpoint,
+    _read_checkpoint,
+)
+from .processing.batch_processor import _scan_log_file, _estimate_batch_bytes
+
 
 @dataclass(frozen=True)
 class IngestStats:
@@ -20,476 +41,6 @@ class IngestStats:
     rows_inserted: int
     rows_skipped: int
 
-
-def _safe_iso_to_datetime(ts: str) -> Optional[dt.datetime]:
-    if not ts:
-        return None
-    try:
-        # Accept timestamps like 2025-06-26T08:42:42.753538Z
-        if ts.endswith("Z"):
-            ts = ts[:-1] + "+00:00"
-        return dt.datetime.fromisoformat(ts)
-    except Exception:
-        return None
-
-
-def _derive_server_id(base_db_dir: Optional[str] = None) -> str:
-    """Derive a stable server id and persist it once per host.
-
-    Resolution order (Stage C):
-    1) LOGDB_SERVER_ID env var (explicit override)
-    2) .server_id file under base_db_dir (if provided)
-    3) Create and persist a new UUID4 into .server_id (if base_db_dir provided)
-    4) Fallback: deterministic UUID5 from hostname+env
-    """
-    # Explicit override is highest priority
-    explicit = os.getenv("LOGDB_SERVER_ID")
-    if explicit:
-        return explicit.strip()
-
-    server_file_path = None
-    if base_db_dir:
-        try:
-            server_file_path = os.path.join(os.path.abspath(base_db_dir), ".server_id")
-            # Read if exists
-            if os.path.isfile(server_file_path):
-                with open(server_file_path, "r", encoding="utf-8") as f:
-                    sid = f.read().strip()
-                    if sid:
-                        return sid
-        except Exception:
-            # Non-fatal: fall through to generation
-            server_file_path = None
-
-    # Generate a new id
-    if server_file_path:
-        try:
-            os.makedirs(os.path.dirname(server_file_path), exist_ok=True)
-            new_id = str(uuid.uuid4())
-            with open(server_file_path, "w", encoding="utf-8") as f:
-                f.write(new_id)
-            return new_id
-        except Exception:
-            pass
-
-    # Last resort: deterministic based on hostname and env
-    env = os.getenv("LOGDB_ENV") or os.getenv("ENV") or "dev"
-    hostname = socket.gethostname()
-    server_uuid = uuid.uuid5(uuid.NAMESPACE_DNS, f"ai-proxy|{hostname}|{env}")
-    return str(server_uuid)
-
-
-def _ensure_servers_row(conn: sqlite3.Connection, server_id: str) -> None:
-    hostname = socket.gethostname()
-    env = os.getenv("LOGDB_ENV") or os.getenv("ENV") or "dev"
-    now = int(dt.datetime.utcnow().timestamp())
-    with conn:
-        conn.execute(
-            """
-            INSERT OR IGNORE INTO servers (server_id, hostname, env, first_seen_ts)
-            VALUES (?, ?, ?, ?)
-            """,
-            (server_id, hostname, env, now),
-        )
-
-
-def _file_sha256(path: str) -> str:
-    h = hashlib.sha256()
-    with open(path, "rb") as f:
-        for chunk in iter(lambda: f.read(65536), b""):
-            h.update(chunk)
-    return h.hexdigest()
-
-
-def _file_prefix_sha256(path: str, upto_bytes: int) -> str:
-    h = hashlib.sha256()
-    read_left = max(0, int(upto_bytes))
-    with open(path, "rb") as f:
-        while read_left > 0:
-            chunk = f.read(min(65536, read_left))
-            if not chunk:
-                break
-            h.update(chunk)
-            read_left -= len(chunk)
-    return h.hexdigest()
-
-
-def _iter_json_blocks(f) -> Iterator[Tuple[int, str]]:
-    """Yield (block_end_pos, json_text) for each rendered JSON block in a log file.
-
-    The log format is: "asctime - level - {\n  ... pretty json ...\n}"
-    JSON can span multiple lines. We detect the first '{' and then
-    track brace depth until it returns to zero.
-    """
-    buffer: Optional[List[str]] = None
-    depth = 0
-    while True:
-        line = f.readline()
-        if not line:
-            # EOF
-            if buffer is not None:
-                # Incomplete trailing block; drop it safely
-                buffer = None
-                depth = 0
-            break
-
-        if buffer is None:
-            brace_idx = line.find("{")
-            if brace_idx == -1:
-                continue
-            buffer = [line[brace_idx:]]
-            depth = buffer[-1].count("{") - buffer[-1].count("}")
-            if depth == 0:
-                json_text = "".join(buffer)
-                buffer = None
-                yield f.tell(), json_text
-            continue
-
-        # We are inside a JSON block
-        buffer.append(line)
-        depth += line.count("{") - line.count("}")
-        if depth <= 0:
-            json_text = "".join(buffer)
-            buffer = None
-            yield f.tell(), json_text
-
-
-def _parse_log_entry(json_text: str) -> Optional[Dict]:
-    try:
-        entry = json.loads(json_text)
-    except Exception:
-        return None
-
-    # Minimal required fields to consider this an API request/response entry
-    if not isinstance(entry, dict):
-        return None
-    if "endpoint" not in entry or "request" not in entry or "response" not in entry:
-        return None
-
-    return entry
-
-
-def _normalize_entry(entry: Dict) -> Optional[Dict]:
-    ts_iso = entry.get("timestamp")
-    if ts_iso is None:
-        return None
-    ts_iso = str(ts_iso)
-    dt_obj = _safe_iso_to_datetime(ts_iso)
-    if dt_obj is None:
-        return None
-    if dt_obj.tzinfo is None:
-        # Assume UTC if tz-naive
-        dt_obj = dt_obj.replace(tzinfo=dt.timezone.utc)
-    epoch_sec = int(dt_obj.timestamp())
-
-    endpoint = str(entry.get("endpoint", ""))
-    if not endpoint:
-        return None
-
-    request_obj = entry.get("request")
-    response_obj = entry.get("response")
-    if request_obj is None or response_obj is None:
-        return None
-
-    try:
-        request_json = json.dumps(request_obj, ensure_ascii=False, sort_keys=True)
-        response_json = json.dumps(response_obj, ensure_ascii=False, sort_keys=True)
-    except Exception:
-        return None
-
-    model_original = None
-    if isinstance(request_obj, dict):
-        model_original = request_obj.get("model")
-    model_mapped = None
-    if isinstance(response_obj, dict):
-        model_mapped = response_obj.get("model")
-
-    # Be tolerant to noisy values
-    status_code_val = entry.get("status_code")
-    try:
-        status_code = int(status_code_val) if status_code_val is not None else None
-    except Exception:
-        status_code = None
-
-    latency_val = entry.get("latency_ms")
-    try:
-        latency_ms = float(latency_val) if latency_val is not None else None
-    except Exception:
-        latency_ms = None
-    api_key_hash = entry.get("api_key_hash")
-
-    return {
-        "ts_iso": ts_iso,
-        "epoch_sec": epoch_sec,
-        "endpoint": endpoint,
-        "model_original": model_original,
-        "model_mapped": model_mapped,
-        "status_code": status_code,
-        "latency_ms": latency_ms,
-        "api_key_hash": api_key_hash,
-        "request_json": request_json,
-        "response_json": response_json,
-        "date": dt_obj.astimezone(dt.timezone.utc).date(),
-    }
-
-
-def _compute_request_id(server_id: str, norm: Dict) -> str:
-    req_sha = hashlib.sha256(norm["request_json"].encode("utf-8")).hexdigest()
-    resp_sha = hashlib.sha256(norm["response_json"].encode("utf-8")).hexdigest()
-    key = f"{server_id}|{norm['ts_iso']}|{norm['endpoint']}|{req_sha}|{resp_sha}"
-    return hashlib.sha256(key.encode("utf-8")).hexdigest()
-
-
-def _upsert_ingest_checkpoint(
-    conn: sqlite3.Connection,
-    source_path: str,
-    sha256_hex: str,
-    bytes_ingested: int,
-    mtime: int,
-) -> None:
-    now = int(dt.datetime.utcnow().timestamp())
-    with conn:
-        conn.execute(
-            """
-            INSERT INTO ingest_sources (source_path, sha256, bytes_ingested, mtime, last_scan_ts)
-            VALUES (?, ?, ?, ?, ?)
-            ON CONFLICT(source_path) DO UPDATE SET
-              sha256=excluded.sha256,
-              bytes_ingested=excluded.bytes_ingested,
-              mtime=excluded.mtime,
-              last_scan_ts=excluded.last_scan_ts
-            """,
-            (source_path, sha256_hex, bytes_ingested, mtime, now),
-        )
-
-
-def _read_checkpoint(
-    conn: sqlite3.Connection, source_path: str
-) -> Tuple[int, int, Optional[str]]:
-    cur = conn.execute(
-        "SELECT bytes_ingested, mtime FROM ingest_sources WHERE source_path= ?",
-        (source_path,),
-    )
-    row = cur.fetchone()
-    if not row:
-        return 0, 0, None
-    # bytes_ingested, mtime, sha256
-    cur2 = conn.execute(
-        "SELECT sha256 FROM ingest_sources WHERE source_path= ?",
-        (source_path,),
-    )
-    row2 = cur2.fetchone()
-    return int(row[0] or 0), int(row[1] or 0), (row2[0] if row2 and row2[0] else None)
-
-
-def _estimate_batch_bytes(batch: List[Tuple]) -> int:
-    total = 0
-    for row in batch:
-        # request_json at index -2, response_json at index -1 for our tuple shape
-        try:
-            req = row[-2]
-            resp = row[-1]
-            total += len(req.encode("utf-8")) + len(resp.encode("utf-8"))
-        except Exception:
-            total += 0
-    # include some overhead per row
-    total += len(batch) * 128
-    return total
-
-
-def _env_int(name: str, default: int) -> int:
-    try:
-        return int(os.getenv(name, str(default)).strip())
-    except Exception:
-        return default
-
-
-def _scan_log_file(
-    source_path: str,
-    base_db_dir: str,
-    since: Optional[dt.date],
-    to: Optional[dt.date],
-    server_id: str,
-) -> Tuple[int, int]:
-    """Scan a single log file and ingest records into partitioned DBs.
-
-    Returns (inserted_rows, skipped_rows).
-    """
-    inserted = 0
-    skipped = 0
-
-    # Use dedicated control DB for checkpoints and server registry
-    control_db = ensure_control_database(base_db_dir)
-    conn = open_connection_with_pragmas(control_db)
-    try:
-        _ensure_servers_row(conn, server_id)
-
-        # Determine resume offset
-        prev_bytes, prev_mtime, prev_sha = _read_checkpoint(conn, source_path)
-        stat = os.stat(source_path)
-        start_pos = 0
-        if prev_bytes > 0 and prev_bytes <= stat.st_size:
-            try:
-                # Validate prefix SHA to ensure file was not rewritten
-                current_prefix_sha = _file_prefix_sha256(source_path, prev_bytes)
-                if prev_sha and current_prefix_sha == prev_sha:
-                    start_pos = prev_bytes
-            except Exception:
-                start_pos = 0
-
-        with open(source_path, "r", encoding="utf-8", errors="ignore") as f:
-            if start_pos:
-                # Seek to resume position. If we are at a line boundary, do not drop the next line.
-                # If we are in the middle of a line, drop the remainder of that line.
-                if start_pos > 0:
-                    f.seek(start_pos - 1)
-                    prev_char = f.read(1)
-                    if prev_char == "\n":
-                        f.seek(start_pos)
-                    else:
-                        f.seek(start_pos)
-                        f.readline()
-                else:
-                    f.seek(start_pos)
-
-            last_good_pos = f.tell()
-            # Prepare per-partition batch buffers and connections
-            conns: Dict[str, sqlite3.Connection] = {}
-            batches: Dict[
-                str,
-                List[
-                    Tuple[
-                        str,
-                        str,
-                        int,
-                        str,
-                        Optional[str],
-                        Optional[str],
-                        Optional[int],
-                        Optional[float],
-                        Optional[str],
-                        str,
-                        str,
-                    ]
-                ],
-            ] = {}
-
-            # Stage I: resource caps
-            max_rows_per_batch = max(1, _env_int("LOGDB_BATCH_ROWS", 500))
-            max_batch_bytes = max(0, _env_int("LOGDB_BATCH_KB", 1024) * 1024)
-            max_memory_bytes = max(0, _env_int("LOGDB_MEMORY_MB", 256) * 1024 * 1024)
-
-            def _flush(db_path: str) -> None:
-                nonlocal inserted
-                batch = batches.get(db_path)
-                if not batch:
-                    return
-                pc = conns[db_path]
-                before = pc.total_changes
-                with pc:
-                    pc.executemany(
-                        """
-                        INSERT OR IGNORE INTO requests (
-                          request_id, server_id, ts, endpoint, model_original, model_mapped,
-                          status_code, latency_ms, api_key_hash, request_json, response_json, dialog_id
-                        ) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, NULL)
-                        """,
-                        batch,
-                    )
-                after = pc.total_changes
-                inserted += max(0, after - before)
-                batches[db_path] = []
-
-            def _current_memory_pressure() -> int:
-                # Approximate current queued memory from batches
-                total = 0
-                for b in batches.values():
-                    if b:
-                        total += _estimate_batch_bytes(b)
-                return total
-
-            for end_pos, json_text in _iter_json_blocks(f):
-                entry = _parse_log_entry(json_text)
-                if not entry:
-                    last_good_pos = end_pos
-                    continue
-
-                norm = _normalize_entry(entry)
-                if not norm:
-                    skipped += 1
-                    last_good_pos = end_pos
-                    continue
-
-                # Date range filter
-                if since and norm["date"] < since:
-                    last_good_pos = end_pos
-                    continue
-                if to and norm["date"] > to:
-                    last_good_pos = end_pos
-                    continue
-
-                # Open partition DB for this record's date
-                db_path = ensure_partition_database(
-                    base_db_dir, norm["date"]
-                )  # creates schema if needed
-                if db_path not in conns:
-                    conns[db_path] = open_connection_with_pragmas(db_path)
-                    _ensure_servers_row(conns[db_path], server_id)
-                    batches[db_path] = []
-
-                request_id = _compute_request_id(server_id, norm)
-                batches[db_path].append(
-                    (
-                        request_id,
-                        server_id,
-                        norm["epoch_sec"],
-                        norm["endpoint"],
-                        norm["model_original"],
-                        norm["model_mapped"],
-                        norm["status_code"],
-                        norm["latency_ms"],
-                        norm["api_key_hash"],
-                        norm["request_json"],
-                        norm["response_json"],
-                    )
-                )
-                # Flush if rows threshold exceeded
-                if len(batches[db_path]) >= max_rows_per_batch:
-                    _flush(db_path)
-                else:
-                    # Flush if bytes threshold exceeded
-                    if (
-                        max_batch_bytes > 0
-                        and _estimate_batch_bytes(batches[db_path]) >= max_batch_bytes
-                    ):
-                        _flush(db_path)
-                # Global memory pressure: flush all if over cap
-                if (
-                    max_memory_bytes > 0
-                    and _current_memory_pressure() >= max_memory_bytes
-                ):
-                    for pth in list(batches.keys()):
-                        _flush(pth)
-
-                last_good_pos = end_pos
-
-            # Flush all remaining batches and close connections
-            for path, _ in list(batches.items()):
-                _flush(path)
-            for pc in conns.values():
-                pc.close()
-
-            # Update checkpoint at last complete block boundary with prefix sha
-            # Compute sha of file prefix up to last_good_pos for robust resume
-            file_prefix_sha = _file_prefix_sha256(source_path, last_good_pos)
-            _upsert_ingest_checkpoint(
-                conn, source_path, file_prefix_sha, last_good_pos, int(stat.st_mtime)
-            )
-
-    finally:
-        conn.close()
-
-    return inserted, skipped
 
 
 def ingest_logs(
